Making Platform, Window and Document Mock available in all tests
