+++ packages/firestore/test/unit/specs/spec_test_runner.ts
@@ -52,7 +52,6 @@ import { JsonObject } from '../../../src/model/field_value';
 import { Mutation } from '../../../src/model/mutation';
 import {
   emptyByteString,
-  Platform,
   PlatformSupport
 } from '../../../src/platform/platform';
 import { Connection, Stream } from '../../../src/remote/connection';
@@ -99,6 +98,7 @@ import {
   SharedClientState,
   WebStorageSharedClientState
 } from '../../../src/local/shared_client_state';
+import { MockPlatform, SharedMockStorage } from '../../util/mock_platform';
 
 class MockConnection implements Connection {
   watchStream: StreamBridge<
@@ -383,7 +383,7 @@ abstract class TestRunner {
   private serializer: JsonProtoSerializer;
 
   constructor(
-    protected readonly platform: TestPlatform,
+    protected readonly platform: MockPlatform,
     private sharedWrites: SharedWriteTracker,
     config: SpecConfig
   ) {
@@ -1137,230 +1137,6 @@ class MemoryTestRunner extends TestRunner {
 }
 
 /**
- * `Document` mock that implements the `visibilitychange` API used by Firestore.
- */
-class MockDocument {
-  private _visibilityState: VisibilityState = 'unloaded';
-  private visibilityListener: EventListener | null;
-
-  get visibilityState(): VisibilityState {
-    return this._visibilityState;
-  }
-
-  addEventListener(type: string, listener: EventListener): void {
-    assert(
-      type === 'visibilitychange',
-      "MockDocument only supports events of type 'visibilitychange'"
-    );
-    this.visibilityListener = listener;
-  }
-
-  removeEventListener(type: string, listener: EventListener): void {
-    if (listener === this.visibilityListener) {
-      this.visibilityListener = null;
-    }
-  }
-
-  raiseVisibilityEvent(visibility: VisibilityState): void {
-    this._visibilityState = visibility;
-    if (this.visibilityListener) {
-      this.visibilityListener(new Event('visibilitychange'));
-    }
-  }
-}
-
-/**
- * `WebStorage` mock that implements the WebStorage behavior for multiple
- * clients. To get a client-specific storage area that implements the WebStorage
- * API, invoke `getStorageArea(storageListener)`.
- */
-class SharedMockStorage {
-  private readonly data = new Map<string, string>();
-  private readonly activeClients: Array<{
-    storageListener: EventListener;
-    storageArea: Storage;
-  }> = [];
-
-  getStorageArea(storageListener: EventListener): Storage {
-    const clientIndex = this.activeClients.length;
-    const self = this;
-
-    const storageArea: Storage = {
-      get length(): number {
-        return self.length;
-      },
-      getItem: (key: string) => this.getItem(key),
-      key: (index: number) => this.key(index),
-      clear: () => this.clear(),
-      removeItem: (key: string) => {
-        const oldValue = this.getItem(key);
-        this.removeItem(key);
-        this.raiseStorageEvent(clientIndex, key, oldValue, null);
-      },
-      setItem: (key: string, value: string) => {
-        const oldValue = this.getItem(key);
-        this.setItem(key, value);
-        this.raiseStorageEvent(clientIndex, key, oldValue, value);
-      }
-    };
-
-    this.activeClients[clientIndex] = { storageListener, storageArea };
-
-    return storageArea;
-  }
-
-  private clear(): void {
-    this.data.clear();
-  }
-
-  private getItem(key: string): string | null {
-    return this.data.get(key);
-  }
-
-  private key(index: number): string | null {
-    return Array.from(this.data.keys())[index];
-  }
-
-  private removeItem(key: string): void {
-    this.data.delete(key);
-  }
-
-  private setItem(key: string, data: string): void {
-    this.data.set(key, data);
-  }
-
-  private get length(): number {
-    return this.data.size;
-  }
-
-  private raiseStorageEvent(
-    sourceClientIndex: number,
-    key: string,
-    oldValue: string | null,
-    newValue: string | null
-  ): void {
-    this.activeClients.forEach((client, index) => {
-      // WebStorage doesn't raise events for writes from the originating client.
-      if (sourceClientIndex === index) {
-        return;
-      }
-
-      client.storageListener({
-        key,
-        oldValue,
-        newValue,
-        storageArea: client.storageArea
-      } as any); // tslint:disable-line:no-any Not mocking entire Event type.
-    });
-  }
-}
-
-/**
- * `Window` mock that implements the event and storage API that is used by
- * Firestore.
- */
-class MockWindow {
-  private readonly storageArea: Storage;
-  private storageListener: EventListener = () => {};
-
-  constructor(sharedMockStorage: SharedMockStorage) {
-    this.storageArea = sharedMockStorage.getStorageArea(event =>
-      this.storageListener(event)
-    );
-  }
-
-  get localStorage(): Storage {
-    return this.storageArea;
-  }
-
-  get indexedDB(): IDBFactory {
-    return window.indexedDB;
-  }
-
-  addEventListener(type: string, listener: EventListener): void {
-    switch (type) {
-      case 'storage':
-        this.storageListener = listener;
-        break;
-      case 'unload':
-        // The spec tests currently do not rely on 'unload' listeners.
-        break;
-      default:
-        fail(`MockWindow doesn't support events of type '${type}'`);
-    }
-  }
-
-  removeEventListener(type: string, listener: EventListener): void {
-    if (type === 'storage') {
-      assert(
-        this.storageListener === listener,
-        "Listener passed to 'removeEventListener' doesn't match the current listener."
-      );
-      this.storageListener = () => {};
-    }
-  }
-}
-
-/**
- * Implementation of `Platform` that allows mocking of `document` and `window`.
- */
-class TestPlatform implements Platform {
-  readonly mockDocument: MockDocument | null = null;
-  readonly mockWindow: MockWindow | null = null;
-
-  constructor(
-    private readonly basePlatform: Platform,
-    private readonly mockStorage: SharedMockStorage
-  ) {
-    this.mockDocument = new MockDocument();
-    this.mockWindow = new MockWindow(this.mockStorage);
-  }
-
-  get document(): Document | null {
-    // tslint:disable-next-line:no-any MockWindow doesn't support full Document interface.
-    return this.mockDocument as any;
-  }
-
-  get window(): Window | null {
-    // tslint:disable-next-line:no-any MockWindow doesn't support full Window interface.
-    return this.mockWindow as any;
-  }
-
-  get base64Available(): boolean {
-    return this.basePlatform.base64Available;
-  }
-
-  get emptyByteString(): ProtoByteString {
-    return this.basePlatform.emptyByteString;
-  }
-
-  raiseVisibilityEvent(visibility: VisibilityState): void {
-    if (this.mockDocument) {
-      this.mockDocument.raiseVisibilityEvent(visibility);
-    }
-  }
-
-  loadConnection(databaseInfo: DatabaseInfo): Promise<Connection> {
-    return this.basePlatform.loadConnection(databaseInfo);
-  }
-
-  newSerializer(databaseId: DatabaseId): JsonProtoSerializer {
-    return this.basePlatform.newSerializer(databaseId);
-  }
-
-  formatJSON(value: AnyJs): string {
-    return this.basePlatform.formatJSON(value);
-  }
-
-  atob(encoded: string): string {
-    return this.basePlatform.atob(encoded);
-  }
-
-  btoa(raw: string): string {
-    return this.basePlatform.btoa(raw);
-  }
-}
-/**
  * Runs the specs using IndexedDbPersistence, the creator must ensure that it is
  * enabled for the platform.
  */
@@ -1416,7 +1192,7 @@ export async function runSpec(
 
   const ensureRunner = async clientIndex => {
     if (!runners[clientIndex]) {
-      const platform = new TestPlatform(
+      const platform = new MockPlatform(
         PlatformSupport.getPlatform(),
         sharedMockStorage
       );
